fix: handle useLayoutEffect warning for SSR
