+++ src/hooks/useIsomorphicLayoutEffect.js
@@ -1,0 +1,4 @@
+import { useLayoutEffect, useEffect } from 'react'
+
+export const useIsomorphicLayoutEffect =
+  typeof window === 'undefined' ? useEffect : useLayoutEffect
